[GAME/MAIN] 버튼 클릭 사운드 추가 & 대사 수정

+++ Snow Globe Cleanup Operation/Assets/Scripts/ChangeScene/StartStoryScene.cs
@@ -1,10 +1,18 @@
+using System.Collections;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
 public class startStoryScene : MonoBehaviour
 {
+    public AudioSource clickSound;
+    private bool isClicked = false;
+
     public void StartStoryScene()
     {
+        if (isClicked) return;
+
+        isClicked = true;
+
         PlayerPrefs.DeleteKey("TimerValue");
         PlayerPrefs.DeleteKey("STAGEA");
         PlayerPrefs.DeleteKey("STAGEB");
@@ -13,6 +21,21 @@ public class startStoryScene : MonoBehaviour
         PlayerPrefs.DeleteKey("STAGEE");
 
         PlayerPrefs.Save();
-        SceneManager.LoadScene("StartStory");
+
+        if (clickSound != null)
+        {
+            clickSound.PlayOneShot(clickSound.clip);
+            StartCoroutine(LoadSceneAfterDelay(clickSound.clip.length));
+        }
+        else
+        {
+            SceneManager.LoadScene("StartStory");
+        }
+    }
+
+    private IEnumerator LoadSceneAfterDelay(float delay)
+    {
+        yield return new WaitForSeconds(delay);
+        SceneManager.LoadScene("StartStory"); 
     }
 }
